+++ MediaPortal/Source/UI/SkinEngine/Skin/default/shaders/lineargradient.fx
@@ -1,10 +1,10 @@
-half4x4  worldViewProj : WORLDVIEWPROJ; // Our world view projection matrix
+float4x4  worldViewProj : WORLDVIEWPROJ; // Our world view projection matrix
 
-half4x4  g_transform;
+float4x4  g_transform;
 texture  g_texture; // Color texture 
-half     g_opacity;
-half2    g_startpoint = {0.0f, 0.0f};
-half2    g_endpoint = {1.0f, 1.0f};
+float     g_opacity;
+float2    g_startpoint = {0.0f, 0.0f};
+float2    g_endpoint = {1.0f, 1.0f};
 
 sampler textureSampler = sampler_state
 {
@@ -17,28 +17,28 @@ sampler textureSampler = sampler_state
 // application to vertex structure
 struct a2v
 {
-  half4 Position  : POSITION0;
-  half2 Texcoord  : TEXCOORD0;  // vertex texture coords 
+  float4 Position  : POSITION0;
+  float2 Texcoord  : TEXCOORD0;  // vertex texture coords 
 };
 
 // vertex shader to pixelshader structure
 struct v2p
 {
-  half4 Position   : POSITION;
-  half2 Texcoord   : TEXCOORD0;
+  float4 Position   : POSITION;
+  float2 Texcoord   : TEXCOORD0;
 };
 
 // pixel shader to frame
 struct p2f
 {
-  half4 Color : COLOR0;
+  float4 Color : COLOR0;
 };
 
-half GetColor(half2 pos)
+float GetColor(float2 pos)
 {
-  half2 vPos = pos-g_startpoint;
-  half2 vDist = g_endpoint-g_startpoint;
-  half dist = dot(vPos, vDist) / dot(vDist, vDist);
+  float2 vPos = pos-g_startpoint;
+  float2 vDist = g_endpoint-g_startpoint;
+  float dist = dot(vPos, vDist) / dot(vDist, vDist);
 
   return dist;
 }
@@ -51,9 +51,9 @@ void renderVertexShader(in a2v IN, out v2p OUT)
 
 void renderPixelShader(in v2p IN, out p2f OUT)
 {
-  half4 pos = half4(IN.Texcoord.x, IN.Texcoord.y, 0, 1);
+  float4 pos = float4(IN.Texcoord.x, IN.Texcoord.y, 0, 1);
   pos = mul(pos, g_transform);
-  half dist = GetColor(half2(pos.x, pos.y));
+  float dist = GetColor(float2(pos.x, pos.y));
   dist = clamp(dist, 0, 0.9999);
   OUT.Color = tex1D(textureSampler, dist);
   OUT.Color[3] *= g_opacity;
